fixed child forker

+++ pinode/app/src/sketchrunner.js
@@ -44,15 +44,14 @@ module.exports = {
   falsecallback: function(){},
 
   stop: function(){
-    //this.spawn_exec("killall mplayer", this.falsecallback);
-    //this.spawn_exec("killall omxplayer.bin", this.falsecallback);
     for (idx in this.child_pids){
       var childpid=this.child_pids[idx];
-      console.log(childpid);
       if (typeof childpid != "function"){
-        console.log( "kill pid: " + childpid );
-        process.kill( childpid );
-        this.child_pids.remove( childpid );
+        if (childpid > 0) {
+          console.log( "kill pid: " + childpid );
+          process.kill( childpid );
+          this.child_pids.remove( childpid );
+        }
       }
     }
 
@@ -109,7 +108,7 @@ module.exports = {
   fork_proc: function(cmd, callback){
     var cp = require('child_process');
     var child = cp.fork( approot + '/bin/child-worker' );
-    child.unref();
+    //child.unref();
     child.on('message', function(msg) {
       switch (msg.type) {
         case "stdout":
